requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
mysqlclient==2.2.0
packaging==23.1
pluggy==1.3.0
PyMySQL==1.1.0
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sudo==1.0.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.3
